services/notification-service/requirements.txt
fastapi[standard]==0.115.5
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
pytest==7.4.0
pytest-asyncio==0.21.0
jinja2==3.1.4
pydantic==2.9.2
pydantic-settings==2.5.2
